+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0 - 2019-10-07]
+
+### Added
+- `EvtxParser::records_json_value()` to allow working with records with a `serde_json::Value`. See `test_into_json_value_records` for an example.
+- `EvtxRecord::into_output`, allowing serializing a record using a user-defined `BinXmlOutput` type.
+
+### Changed
+
+- `SerializedEvtxRecord` is now generic over it's `data`, allowing a simplified `BinXmlOutput` trait.
+
 ## [0.4.2 - 2019-09-05]
 
 ### Added